requirements.txt
pynvml>=11.5.0
nvidia-ml-py>=12.535.108
torch>=2.1.0
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0
pyyaml>=6.0
prometheus-client>=0.19.0
requests>=2.31.0
aiohttp>=3.9.0
asyncio-throttle>=1.0.2
jinja2>=3.1.2
rich>=13.7.0
click>=8.1.7
tabulate>=0.9.0
matplotlib>=3.8.0
seaborn>=0.13.0
psutil>=5.9.0
colorlog>=6.8.0
jsonschema>=4.20.0
tqdm>=4.66.0
# vLLM (install separately with: pip install vllm)
# vllm>=0.3.0
